+++ src/main/java/com/mrporter/pomangam/client/controllers/advertisement/AdvertisementController.java
@@ -25,7 +25,7 @@ public class AdvertisementController {
 
     @GetMapping("/{idx}")
     public ResponseEntity<?> getByIdx(
-            @PathVariable(value = "idx", required = true) Integer idx
+            @PathVariable(value = "idx", required = true) Long idx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
