+++ src/main/java/com/mrporter/pomangam/client/controllers/ordertime/OrderTimeController.java
@@ -13,32 +13,32 @@ public class OrderTimeController {
 
     @GetMapping("/dsites/{didx}/ordertimes")
     public ResponseEntity<?> getByIdxDeliverySite(
-            @PathVariable(value = "didx", required = true) Integer didx
+            @PathVariable(value = "didx", required = true) Long didx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
 
     @GetMapping("/dsites/{didx}/ordertimes/{idx}")
     public ResponseEntity<?> getByIdxAndIdxDeliverySite(
-            @PathVariable(value = "didx", required = true) Integer didx,
-            @PathVariable(value = "idx", required = true) Integer idx
+            @PathVariable(value = "didx", required = true) Long didx,
+            @PathVariable(value = "idx", required = true) Long idx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
 
     @GetMapping("/dsites/{didx}/stores/{sidx}/ordertimes")
     public ResponseEntity<?> getByIdxDeliverySiteAndIdxStore(
-            @PathVariable(value = "didx", required = true) Integer didx,
-            @PathVariable(value = "sidx", required = true) Integer sidx
+            @PathVariable(value = "didx", required = true) Long didx,
+            @PathVariable(value = "sidx", required = true) Long sidx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
 
     @GetMapping("/dsites/{didx}/stores/{sidx}/ordertimes/{idx}")
     public ResponseEntity<?> getByIdxAndIdxDeliverySiteAndIdxStore(
-            @PathVariable(value = "didx", required = true) Integer didx,
-            @PathVariable(value = "sidx", required = true) Integer sidx,
-            @PathVariable(value = "idx", required = true) Integer idx
+            @PathVariable(value = "didx", required = true) Long didx,
+            @PathVariable(value = "sidx", required = true) Long sidx,
+            @PathVariable(value = "idx", required = true) Long idx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
