+++ src/main/java/com/mrporter/pomangam/client/controllers/store/review/StoreReviewController.java
@@ -18,8 +18,8 @@ public class StoreReviewController {
 
     @GetMapping
     public ResponseEntity<?> get(
-            @PathVariable(value = "didx", required = true) Integer didx,
-            @PathVariable(value = "sidx", required = true) Integer sidx,
+            @PathVariable(value = "didx", required = true) Long didx,
+            @PathVariable(value = "sidx", required = true) Long sidx,
             @PageableDefault(sort = {"idx"}, direction = Sort.Direction.DESC, size = 10) Pageable pageable
     ) {
         return new ResponseEntity(HttpStatus.OK);
@@ -27,17 +27,17 @@ public class StoreReviewController {
 
     @GetMapping("/{idx}")
     public ResponseEntity<?> getByIdx(
-            @PathVariable(value = "didx", required = true) Integer didx,
-            @PathVariable(value = "sidx", required = true) Integer sidx,
-            @PathVariable(value = "idx", required = true) Integer idx
+            @PathVariable(value = "didx", required = true) Long didx,
+            @PathVariable(value = "sidx", required = true) Long sidx,
+            @PathVariable(value = "idx", required = true) Long idx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
 
     @GetMapping("/search/count")
     public ResponseEntity<?> searchCount(
-            @PathVariable(value = "didx", required = true) Integer didx,
-            @PathVariable(value = "sidx", required = true) Integer sidx
+            @PathVariable(value = "didx", required = true) Long didx,
+            @PathVariable(value = "sidx", required = true) Long sidx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
