+++ src/main/java/com/mrporter/pomangam/client/repositories/_bases/CommonMapJpaRepository.java
@@ -9,10 +9,10 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import java.util.List;
 
 @RepositoryRestResource(exported = false)
-public interface CommonMapJpaRepository extends JpaRepository<CommonMap, Integer> {
+public interface CommonMapJpaRepository extends JpaRepository<CommonMap, Long> {
 
     List<CommonMap> findByKey(@Param("key") String key);
 
     @Query(value = "SELECT * FROM common_map_tbl WHERE idx = :idx", nativeQuery = true)
-    CommonMap findByIdx(@Param("idx") Integer idx);
+    CommonMap findByIdx(@Param("idx") Long idx);
 }
