+++ src/main/java/com/mrporter/pomangam/client/repositories/_bases/PolicyJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(exported = false)
-public interface PolicyJpaRepository extends JpaRepository<Policy, Integer> {
+public interface PolicyJpaRepository extends JpaRepository<Policy, Long> {
     Policy getByPolicyName(@Param("policyName") String policyName);
 }
