+++ src/main/java/com/mrporter/pomangam/client/repositories/deliverysite/DeliverySiteJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(exported = false)
-public interface DeliverySiteJpaRepository extends JpaRepository<DeliverySite, Integer> {
+public interface DeliverySiteJpaRepository extends JpaRepository<DeliverySite, Long> {
 
 }
